Rename plan to create
